Clean up directory structure to minimize excess empty folders
The previous directory structure was causing the Typescript compiler
to create a lot of empty folders in other to preserve the relative
paths between source files. This created a very cluttered and confusing
directory, so I've decided that any files shared between the client
and server will exist within the server directory.