Modify switch component to expose less props and contain documentation
Previously, it was possible to pass in the desired classes to be applied
to a switch component and its options before and after interaction. This
has been modified to simplify the prop list and suggest that a default
behavior is present.

Instead, the switch will apply a strict set of classes to all elements.
If desired, these can be overridden by setting the id prop on the switch
and creating new style rules in the stylesheet.